+++ NEWS.md
@@ -1,3 +1,7 @@
+# ddml (development version)
+
+- Adds support for the average treatment effect on the treated estimator.
+
 # ddml 0.1.0
 
 * Initial CRAN submission.